Integration testing complete with minor bug fixes
Fixed bug in Data module that showed up when the JSON file is empty. Fixed bug in User module so no errors appear if name and password aren't entered as strings by the user. Fixed a bug in Modes module so that a warning message is logged if the wrong data type is entered in any public function.